+++ test/auth/auth_module_test.dart
@@ -1,4 +1,4 @@
-import 'package:test/test.dart';
+import 'package:flutter_test/flutter_test.dart';
 import 'package:synheart_core/src/auth/auth_module.dart';
 
 void main() {